ci: add testnet bin to release process
The `testnet` binary is added to the release process. I initially wanted to have it as a separate
release, but this won't work because our installation process is relying on being able to get the
latest version of the client and node from what Github has marked as the latest release for the
repository. Having two separate types of releases would confuse this situation.

This also refactors the Makefile targets to a Justfile. Just is a modern alternative to Make
implemented in Rust that has features that greatly help reduce duplication.

Everything that was done with Github Actions is now done with shell scripts in the Justfile. The
reason for this is so that the process can be run manually from the command line when something goes
wrong with a release and it needs to be completed manually. Quite honestly I think it is actually
easier to understand like this anyway.