fix(node): check when handling msg if is from member
The previous check for "is from me" was failing due to
a local 0.0.0.0 representation in testnets vs 127 from other nodes...

This was leading to looping.

But only checking if it was forwarded from ourselves would mean if
an elder was a storage node, we'dsee and forward on the same message
several times (as each elder sends it on to us, and its not from us, so
we forward to the adults in question etc).

This check solves both situations