chore: re-enable tracing env-filter feature
The `sn_testnet` crate still requires this feature or it won't compile.

The release commit that bumped the versions after this change was reverted because the `sn_testnet`
crate couldn't be published. This is the first crate in the chain, so we can just revert the whole
release commit and the next time we merge we will get a version bump without the `sn_testnet`
change.